# api-spec.yaml
openapi: 3.0.3
info:
  title: Pet Clinic API
  description: API para gerenciamento de Users, Clients, Owners, Pets e Appointments com máquina de estados.
  version: 1.0.0
  contact:
    name: Construção de Software PUCRS
    email: contato@exemplo.com
servers:
  - url: http://localhost:3000/api/v1
    description: Servidor de desenvolvimento local
  - url: https://seu-app.herokuapp.com/api/v1
    description: Servidor de produção
security:
  - bearerAuth: []
...
paths:
...
  /users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Users]
      summary: Buscar usuário por ID
      operationId: getUserById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Users]
      summary: Atualizar usuário
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Usuário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Users]
      summary: Deletar usuário
      operationId: deleteUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Usuário deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /clients/{clientId}:
    parameters:
      - in: path
        name: clientId
        required: true
        schema:
          type: string
    get:
      tags: [Clients]
      summary: Buscar client por ID
      operationId: getClientById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Client encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Clients]
      summary: Atualizar client
      operationId: updateClient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientRequest'
      responses:
        '200':
          description: Client atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Clients]
      summary: Deletar client
      operationId: deleteClient
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Client deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /owners/{ownerId}:
    parameters:
      - in: path
        name: ownerId
        required: true
        schema:
          type: string
    get:
      tags: [Owners]
      summary: Buscar owner por ID
      operationId: getOwnerById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Owner encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Owners]
      summary: Atualizar owner
      operationId: updateOwner
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOwnerRequest'
      responses:
        '200':
          description: Owner atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Owners]
      summary: Deletar owner
      operationId: deleteOwner
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Owner deletado com sucesso
        '409':
          description: Não é permitido deletar owner com pets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /pets:
    get:
      tags: [Pets]
      summary: Listar pets
      operationId: listPets
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: ownerId
          schema:
            type: string
          description: Filtrar por ownerId
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de pets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponsePets'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Pets]
      summary: Criar pet
      operationId: createPet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePetRequest'
      responses:
        '201':
          description: Pet criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Owner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          $ref: '#/components/responses/Conflict'
  /pets/{petId}:
    parameters:
      - in: path
        name: petId
        required: true
        schema:
          type: string
    get:
      tags: [Pets]
      summary: Buscar pet por ID
      operationId: getPetById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pet encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Pets]
      summary: Atualizar pet
      operationId: updatePet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePetRequest'
      responses:
        '200':
          description: Pet atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Pets]
      summary: Deletar pet
      operationId: deletePet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pet deletado com sucesso
        '409':
          description: Não é permitido deletar pet com appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Acesso não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unauthorized:
              value:
                error: "Unauthorized"
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid:
              value:
                error: "Invalid input data"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            not_found:
              value:
                error: "Not found"
    Conflict:
      description: Conflito (p.ex. recurso duplicado ou regra de negócio)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            conflict:
              value:
                error: "Conflict"
    InternalError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            server_error:
              value:
                error: "internal"
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid input data"
        details:
          type: string
          example: "Campo X é obrigatório"
    PaginatedResponseBase:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          example: 100
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0
        hasMore:
          type: boolean
          example: true

    # ------------------------
    # Users (admin)
    # ------------------------
    User:
      type: object
      required:
        - id
        - name
        - email
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "9ea6f22e-2c39-466a-b65f-e8fc1ae9aa01"
        name:
          type: string
          example: "João Silva"
        email:
          type: string
          format: email
          example: "joao@example.com"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
...
    Client:
      type: object
      required:
        - clientId
        - name
        - email
        - createdAt
      properties:
        clientId:
          type: string
          example: "c1a5e7b8-1234-4a2b-9f2a-0b1c2d3e4f5a"
        name:
          type: string
          example: "Empresa X"
        email:
          type: string
          format: email
          example: "contato@empresa.com"
        phone:
          type: string
          example: "+55 51 99999-9999"
        createdAt:
          type: string
          format: date-time
...
    UpdateClientRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
...
    Owner:
      type: object
      required:
        - ownerId
        - name
        - email
        - phone
        - createdAt
      properties:
        ownerId:
          type: string
          example: "o1a2b3c4-1234-5678-9abc-def012345678"
        name:
          type: string
          example: "Ana Laura Lopes"
        email:
          type: string
          format: email
          example: "ana@example.com"
        phone:
          type: string
          example: "+55 51 99999-9999"
        address:
          type: string
          example: "Rua Exemplo, 123"
        createdAt:
          type: string
          format: date-time
...
    UpdateOwnerRequest:
      type: object
      properties:
        name: { type: string }
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
...
    Pet:
      type: object
      required:
        - petId
        - ownerId
        - name
        - species
        - createdAt
      properties:
        petId:
          type: string
          example: "p1a2b3c4-9876-5432-1abc-def012345678"
        ownerId:
          type: string
          example: "o1a2b3c4-1234-5678-9abc-def012345678"
        name:
          type: string
          example: "Rex"
        species:
          type: string
          example: "cachorro"
        breed:
          type: string
          example: "Labrador"
        birthDate:
          type: string
          format: date
          example: "2018-05-20"
        weight:
          type: number
          example: 12.5
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
    CreatePetRequest:
      type: object
      required:
        - ownerId
        - name
        - species
      properties:
        ownerId:
          type: string
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        birthDate:
          type: string
          format: date
        weight:
          type: number
        notes:
          type: string
    UpdatePetRequest:
      type: object
      properties:
        ownerId:
          type: string
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        birthDate:
          type: string
          format: date
        weight:
          type: number
        notes:
          type: string
    PaginatedResponsePets:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponseBase'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Pet'

    # ------------------------
    # Appointments and state machine
    # ------------------------
...